LayoutTests/inspector/storage-panel-dom-storage-empty-expected.txt

Test that storage panel is present and that it contains no data for local and session DOM storages.

Populated local and session storage
Did show: Local storage
Did show: Session storage
Local storage content: 
KeyValue pairs: 
Session storage content: 
KeyValue pairs: 
DONE
